+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/discord/DiscordUtils.java
@@ -109,6 +109,10 @@ public class DiscordUtils {
      */
     public String CLIENT_ID;
     /**
+     * Whether to register this application as an application with discord
+     */
+    public boolean AUTO_REGISTER;
+    /**
      * The Current Starting Unix Timestamp from Epoch, used for Elapsed Time
      */
     public long START_TIMESTAMP;
@@ -203,7 +207,7 @@ public class DiscordUtils {
             }
 
             // Create IPC Instance and Listener and Make a Connection if possible
-            ipcInstance = new IPCClient(Long.parseLong(CLIENT_ID), ModUtils.IS_DEV, ModUtils.IS_VERBOSE, true, CLIENT_ID);
+            ipcInstance = new IPCClient(Long.parseLong(CLIENT_ID), ModUtils.IS_DEV, ModUtils.IS_VERBOSE, AUTO_REGISTER, CLIENT_ID);
             ipcInstance.setListener(new ModIPCListener());
             ipcInstance.connect();